+++ src/commands/exit_command.rs
@@ -1,0 +1,9 @@
+
+use crate::console::{log,LogLevel};
+use crate::employee::EmployeeFile;
+use crate::files::DataFiles;
+
+pub fn exit_command(_: Vec<String>, _: &DataFiles, _: &mut EmployeeFile) {
+    log(LogLevel::INFO, format!("Thank you for using the ID system!"));
+    std::process::exit(0);
+}
